Add parents to git log message
There are scenarios where automated systems may build artifacts from
merge commits. For example, when a GitHub Action is triggered by a PR,
GitHub creates a merge commit that merges the PR branch into the target
branch. The hash of this merge commit is useless, as it is short-lived
and does not refer to any commit in the repository; instead, this commit
is local to GitHub. By also logging the parent commit IDs, one can
determine exactly what code is running.

References:
- https://github.com/orgs/community/discussions/26325
- https://docs.github.com/en/actions/reference/workflows-and-actions/events-that-trigger-workflows#pull_request

PR #7